1st attempt at ML event finder
The general structure of the neural network is now established! Everything from here forward will be to optimize the inputs and outputs to improve the model accuracy. A preliminary run creates a model accuracy of ~87% with a loss of ~0.32. These values are rather susceptible to change depending on the randomized selection of non-reconnection events the algorithm performs in its initiation.